requirements.txt
python-dotenv==1.0.0
requests==2.27.1
terminaltables==3.1.10
